stats237_course_v1_3_0/requirements.txt
# Core
pypdf>=5.0.0
python-dateutil>=2.9.0
pandas>=2.2.0
numpy>=2.0.0
scipy>=1.12.0
pytest>=8.0.0
matplotlib>=3.8.0
hypothesis>=6.100.0

# API
fastapi>=0.110.0
uvicorn>=0.29.0
httpx>=0.27.0

# Book build (HTML) — install on demand
mkdocs>=1.6.0
mkdocs-material>=9.5.0
mkdocs-jupyter>=0.24.0
pymdown-extensions>=10.8
jupyter>=1.0.0
nbconvert>=7.16.0

# PDF export (fallback uses ReportLab; optional HTML→PDF tooling can be added later)
reportlab>=4.0.0

# Reading Edition (handwritten notes)
pymupdf>=1.24.0
pillow>=10.0.0

pyyaml>=6.0.0
